+++ magic_migrate/src/declarative_macros.rs
